place all files in final place and fix imports. All components have errors because they are not typescript
