+++ client/src/lib/components/FullInfo.svelte
