+++ client/src/lib/components/MovieList.svelte
