+++ client/src/lib/components/NoResultsGif.svelte
