+++ client/src/lib/components/PerformanceInfoPopup.svelte
