+++ client/src/lib/components/Seating.svelte
