+++ client/src/lib/components/card_v2/TitleAndStats.svelte
