+++ client/src/lib/components/performanceFilterCard.svelte
