+++ client/src/routes/+page.ts
@@ -1,0 +1,1 @@
+export const ssr = false; // Disable SSR for this route (because of accessing the 'window' object)
