--- client/src/src/main.js
@@ -1,10 +1,0 @@
-import App from "./App.svelte";
-
-const app = new App({
-  target: document.body,
-  props: {
-    //name: 'world'
-  },
-});
-
-export default app;
